BUG: Fix mask assignment in masked_where to use .mask property.
This commit fixes a bug in numpy.ma.masked_where when it is passed a
structured array.  masked_where was inadvertently setting the _mask
field of a structured array to a non-structured array, so we would end
up with a masked structured array where x.data was structured, but
x.mask was not.  This led to troubles for methods such as __getitem__,
__repr__, and tolist.  By writing to the property .mask instead of
._mask, this problem is fixed.

Closes #2972.