MAINT: Refactor numpy.ma.where.
Use np._NoValue as default values of 'x' and 'y' in signature. That
allows None values to be used for compatibility with the unmasked
version of where and makes detection of the case when neither is
given easier.

Use properties *.data and *.mask instead of *._data and *._mask. That
may avoid some subtle problems.

Minor style cleanups.