BUG: Fixed slicing of chararrays on Python 3.
When taking a slice of a chararray it was calling the rstrip() method
on the resulting slice, resulting in a new array rather than a view
of the original.  This was an unintended consequence of the sq_slice
member of the tp_as_sequence mapping being ignored in Python 3, so
that slice lookups go directly through __getitem__.

Fix test_expandtabs to not make the assumption that rstrip() will be applied twice to a value when using T[x][y] style indexing.