MAINT: remove legacy monkeypatching of GzipFile
I'm not sure exactly when GzipFile.seek started supporting the whence=
argument by default -- sometime around python 2.5 from the looks of
http://bugs.python.org/issue1355023. But in any case it was definitely
there by 2.6, which is now the earliest version we support, so there's
no longer any need to monkeypatch it in. This also fixes an error in
python 3.5b2, which I haven't bothered to track down further because
these are the wages of monkeypatching.