BUG: Fixed downcasting with boolean indexing.
Fixed downcasting complex types to a real type using boolean indices which
previously caused a segfault.  Acquiring the appropriate dtype transfer
function raises a warning that requires holding the GIL, which had been
released in a previous revision.  Now the GIL is held while the transfer
function is acquired.

This fixes issue #5896.