DEV: Fix dates on Travis's time as leader
The 2001 date I had originally was based on the beginning of NumPy's git
history -- but further research suggests that this actually runs back
into Numeric times, before the NumPy project was founded.

New 2005 date is based on Fernando's recollection of the fateful meeting
that led to NumPy's creation, and is consistent with Wikipedia's claim
that 2006 was the first release.

If this is wrong please let me know :-)