BUG: immutable _arraymethod function in ma.core
Replace the _arraymethod class in ma.core with a function factory which
returns class method wrappers around basic array methods. These methods are
bound to the MaskedArray instance and are immutable.

Previously _arraymethod was a class which would incorrectly operate on the
MaskedArray object which last accessed the particular named function.

closes #5247