BUG: npy_PyFile_Dup2 - fix PyFile_AsFile failing on io style classes
This bug only manifests itself in the Python 2 code path.

Falls back to the current Python 3 code path also on Python 2 because
the Python 3 code path is written precisely to handle this situation.

Also fix tests, and clarify in the documentation that the current
implementation requires the stream to be seekable.