TST: Clean up travis-test and make it work with current travis ci.
Travis ci is migrating to GCI and the 32 bit tests broke in the process.
This cleans up the tools/travis-test script, fixes it for current
travis, and changes the 32 bit test to use python 2.7 in order to turn
up errors involving python long integers.

In preparation for dropping Python 3.2 and 3.3, the USE_DEBUG test is
run in the travis ci trusty beta so that python3 defaults to 3.4.