ENH: Add context manager `temppath` to manage a temporary file.
Context manager intended for use when the same temporary file needs to
be opened and closed more than once. The context manager creates the
file, closes it, and returns the path to the file. On exit from the
context block the file is removed. The file should be closed before
exiting the context as an error will be raised on windows if not.

Also fix up the `tempdir` context manager to deal with exceptions.

Tests are added for both `temppath` and `tempdir`.