BUG: ignore exceptions in numpy/tests/test_scripts.py/test_f2p
The test was checking whether the f2py script was installed as either of
two names, but was only catching OSError, so the second check was
skipped if the first failed for another reason. The caused the
runtests.py script to fail it does not install the script as f2py but
rather with the python version appended.